nextintranet_backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.2
django-colorfield==0.11.0
django-constance==4.1.3
django-cors-headers==4.5.0
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
graphene==3.4
graphene-django==3.2.2
# graphql-core==3.2.5
graphql-relay
django-graphql-jwt
django-graphql-auth
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.0.0
promise==2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
redis==5.2.1
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.2
psycopg2
fpdf2
# pylibdmtx
# pyStrich
treepoem
django-crispy-forms
crispy-bootstrap5
django-select2
django-htmx
django-tables2
django-vanilla-views
django-gravatar
django-money
django-location-field
django-formtools
requests
drf-spectacular[sidecar]
